creates network provider hook
